unit tests for engine and battery
