+++ .gitignore
@@ -132,3 +132,4 @@ dmypy.json
 .pyre/
 
 .trunk
+.vscode
