+++ engine/sternman_engine.py
@@ -4,5 +4,5 @@ class SternmanEngine(Engine):
     def __init__(self, warning_light_on):
         self.__warning_light_on = warning_light_on
 
-    def engine_should_be_serviced(self):
+    def needs_service(self):
         return self.__warning_light_on
