+++ test/test_battery.py
@@ -1,0 +1,34 @@
+import unittest
+from datetime import datetime
+
+from battery.nubbin_battery import NubbinBattery
+from battery.spindler_battery import SpindlerBattery
+
+class TestNubbinBattery(unittest.TestCase):
+    def test_battery_should_be_serviced(self):
+        today = datetime.today().date()
+        last_service_date = today.replace(year=today.year - 5)
+        battery = NubbinBattery(today, last_service_date)
+        self.assertTrue(battery.needs_service())
+
+    def test_battery_should_not_be_serviced(self):
+        today = datetime.today().date()
+        last_service_date = today.replace(year=today.year - 3)
+        battery = NubbinBattery(today, last_service_date)
+        self.assertFalse(battery.needs_service())
+
+class TestSpindlerBattery(unittest.TestCase):
+    def test_battery_should_be_serviced(self):
+        today = datetime.today().date()
+        last_service_date = today.replace(year=today.year - 3)
+        battery = SpindlerBattery(today, last_service_date)
+        self.assertTrue(battery.needs_service())
+
+    def test_battery_should_not_be_serviced(self):
+        today = datetime.today().date()
+        last_service_date = today.replace(year=today.year - 1)
+        battery = SpindlerBattery(today, last_service_date)
+        self.assertFalse(battery.needs_service())
+
+if __name__ == '__main__':
+    unittest.main()
