+++ test/test_engine.py
@@ -1,0 +1,44 @@
+import unittest
+from engine.capulet_engine import CapuletEngine
+from engine.sternman_engine import SternmanEngine
+from engine.willoughby_engine import WilloughbyEngine
+
+class TestCapuletEngine(unittest.TestCase):
+    def test_engine_should_be_serviced(self):
+        current_mileage = 30001
+        last_service_mileage = 0
+        engine = CapuletEngine(current_mileage, last_service_mileage)
+        self.assertTrue(engine.needs_service())
+
+    def test_engine_should_not_be_serviced(self):
+        current_mileage = 30000
+        last_service_mileage = 0
+        engine = CapuletEngine(current_mileage, last_service_mileage)
+        self.assertFalse(engine.needs_service())
+
+class TestSternmanEngine(unittest.TestCase):
+    def test_engine_should_be_serviced(self):
+        warning_light_on = True
+        engine = SternmanEngine(warning_light_on)
+        self.assertTrue(engine.needs_service())
+
+    def test_engine_should_not_be_serviced(self):
+        warning_light_on = False
+        engine = SternmanEngine(warning_light_on)
+        self.assertFalse(engine.needs_service())
+
+class TestWilloughbyEngine(unittest.TestCase):
+    def test_engine_should_be_serviced(self):
+        current_mileage = 60001
+        last_service_mileage = 0
+        engine = WilloughbyEngine(current_mileage, last_service_mileage)
+        self.assertTrue(engine.needs_service())
+
+    def test_engine_should_not_be_serviced(self):
+        current_mileage = 60000
+        last_service_mileage = 0
+        engine = WilloughbyEngine(current_mileage, last_service_mileage)
+        self.assertFalse(engine.needs_service())
+
+if __name__ == '__main__':
+    unittest.main()
